requirements.txt
fastapi[standard]
alembic
python-dotenv
sqlmodel
psycopg2-binary
langchain 
langchain_community 
langchain_openai 
langchain_chroma
pytest
pytest-asyncio
python-multipart
pypdf
streamlit